review2, moved logger service into common module
