+++ apps/api/src/approval/controller/approval.controller.spec.ts
@@ -15,14 +15,14 @@ import { RedisClientType } from 'redis'
 import { mockDeep } from 'jest-mock-extended'
 import { ProviderModule } from '../../provider/provider.module'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('ApprovalController', () => {
   let controller: ApprovalController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [ProviderModule, LoggerModule],
+      imports: [ProviderModule, CommonModule],
       controllers: [ApprovalController],
       providers: [
         ApprovalService,
