+++ apps/api/src/environment/service/environment.service.spec.ts
@@ -3,14 +3,14 @@ import { EnvironmentService } from './environment.service'
 import { PrismaService } from '../../prisma/prisma.service'
 import { mockDeep } from 'jest-mock-extended'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('EnvironmentService', () => {
   let service: EnvironmentService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [LoggerModule],
+      imports: [CommonModule],
       providers: [EnvironmentService, PrismaService, AuthorityCheckerService]
     })
       .overrideProvider(PrismaService)
