+++ apps/api/src/event/controller/event.controller.spec.ts
@@ -3,14 +3,14 @@ import { EventController } from './event.controller'
 import { EventService } from '../service/event.service'
 import { PrismaService } from '../../prisma/prisma.service'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('EventController', () => {
   let controller: EventController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [LoggerModule],
+      imports: [CommonModule],
       controllers: [EventController],
       providers: [EventService, PrismaService, AuthorityCheckerService]
     }).compile()
