+++ apps/api/src/event/service/event.service.spec.ts
@@ -2,14 +2,14 @@ import { Test, TestingModule } from '@nestjs/testing'
 import { EventService } from './event.service'
 import { PrismaService } from '../../prisma/prisma.service'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('EventService', () => {
   let service: EventService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [LoggerModule],
+      imports: [CommonModule],
       providers: [EventService, PrismaService, AuthorityCheckerService]
     }).compile()
