+++ apps/api/src/event/service/event.service.ts
@@ -7,7 +7,7 @@ import { AuthorityCheckerService } from '../../common/authority-checker.service'
 export class EventService {
   constructor(
     private readonly prisma: PrismaService,
-    public authorityCheckerService: AuthorityCheckerService
+    private readonly authorityCheckerService: AuthorityCheckerService
   ) {}
 
   async getEvents(
