--- apps/api/src/logger/logger.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common'
-import { CustomLoggerService } from './logger.service'
-
-@Module({
-  providers: [CustomLoggerService],
-  exports: [CustomLoggerService]
-})
-export class LoggerModule {}
