+++ apps/api/src/project/service/project.service.spec.ts
@@ -5,14 +5,14 @@ import { MAIL_SERVICE } from '../../mail/services/interface.service'
 import { PrismaService } from '../../prisma/prisma.service'
 import { mockDeep } from 'jest-mock-extended'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('ProjectService', () => {
   let service: ProjectService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [LoggerModule],
+      imports: [CommonModule],
       providers: [
         ProjectService,
         PrismaService,
