+++ apps/api/src/secret/service/secret.service.ts
@@ -51,7 +51,7 @@ export class SecretService {
     readonly redisClient: {
       publisher: RedisClientType
     },
-    public authorityCheckerService: AuthorityCheckerService
+    private readonly authorityCheckerService: AuthorityCheckerService
   ) {
     this.redis = redisClient.publisher
   }
