+++ apps/api/src/workspace-role/service/workspace-role.service.spec.ts
@@ -4,14 +4,14 @@ import { PrismaService } from '../../prisma/prisma.service'
 import { MAIL_SERVICE } from '../../mail/services/interface.service'
 import { MockMailService } from '../../mail/services/mock.service'
 import { AuthorityCheckerService } from '../../common/authority-checker.service'
-import { LoggerModule } from '../../logger/logger.module'
+import { CommonModule } from '../../common/common.module'
 
 describe('WorkspaceRoleService', () => {
   let service: WorkspaceRoleService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [LoggerModule],
+      imports: [CommonModule],
       providers: [
         WorkspaceRoleService,
         PrismaService,
